add utilities.py, which contains some functions possibly needed by gamemode events
this also fixes it so priest death events finally happen
